requirements.txt
PyMuPDF
Pillow
numpy<2
easyocr
pyobjc-framework-Vision
pyobjc-framework-Quartz
google-generativeai
config